simplify simple story
